+++ include/timetable/bounds.h
@@ -19,11 +19,8 @@ namespace Timetable {
   using bounds_t          = bounds_template<visit_list::const_iterator>;
   using reverse_bounds_t  = bounds_template<visit_list::const_reverse_iterator>;
 
-  reverse_bounds_t reverse(bounds_t bounds) {
+  template<typename BoundsType>
+  reverse_bounds_t reverse(BoundsType bounds) {
     return { std::make_reverse_iterator(bounds.end()), std::make_reverse_iterator(bounds.begin()) };
-  };
-
-  bounds_t reverse(reverse_bounds_t bounds) {
-    return { bounds.end().base(), bounds.begin().base() };
   };
 }
