+++ include/timetable/timetable.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <iostream>
 #include <map>
 #include <unordered_map>
 
@@ -56,10 +57,9 @@ namespace Timetable {
       ////
 
       reverse_bounds_t visits_before(const visit_list_key key) const {
-        auto lower_bound = visits.lower_bound(key.station_lower_bound());
         auto upper_bound = visits.upper_bound(key);
-        bounds_t bounds{ lower_bound, upper_bound };
-        return reverse(bounds);
+        auto lower_bound = visits.lower_bound(key.station_lower_bound());
+        return reverse(bounds_t{ lower_bound, upper_bound });
       };
 
       bounds_t visits_between(const visit_list_key key1, const visit_list_key key2) const {
